Set disposal method to "background"
The "do not dispose" disposal method replaces transparent pixels with
the corresponding pixel from the previous frame - in effect, drawing the
new frame on top of the old frame. This is useful as a poor-man's delta
encoding scheme, but it makes it impossible to replace a pixel with a
transparent pixel.

The "background" disposal method replaces transparent pixels with the
background color (which this module currently sets to color 0).

The default disposal method is undefined, so clients are free to do
whatever they want. Firefox at least seems to treat it the same as "do
not dispose".